- Thay đổi back end - Validate dữ liệu

+++ src/main/java/com/example/event_management/DTO/EventDTO.java
@@ -8,9 +8,9 @@ import java.sql.Time;
 @Data
 public class EventDTO {
 
-    private int event_id ;
-    private int event_admin_id ;
-    private int event_speaker_id ;
+    private long event_id ;
+    private long event_admin_id ;
+    private long event_speaker_id ;
     private String event_name ;
     private Date event_date ;
     private Time event_starting_time ;
