+++ src/main/java/com/example/event_management/repository/AdminRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface AdminRepository extends JpaRepository <AdminEntity , Integer> {
+public interface AdminRepository extends JpaRepository <AdminEntity , Long> {
 
     AdminEntity findAdminEntityByAdminAccountName(String name);
