+++ src/main/java/com/example/event_management/repository/EventRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 
-public interface EventRepository extends JpaRepository <EventEntity, Integer> {
+public interface EventRepository extends JpaRepository <EventEntity, Long> {
