+++ src/main/java/com/example/event_management/repository/RegisterRepository.java
@@ -4,7 +4,7 @@ import com.example.event_management.entity.AdminEntity;
 import com.example.event_management.entity.RegisterEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RegisterRepository extends JpaRepository <RegisterEntity, Integer> {
+public interface RegisterRepository extends JpaRepository <RegisterEntity, Long> {
 
     RegisterEntity findRegisterEntityByRegisterAccountName(String name) ;
